LayoutTests/svg/dynamic-updates/SVGFilterElement-svgdom-filterResY-prop-expected.txt

CONSOLE MESSAGE: line 44: TypeError: 'undefined' is not an object (evaluating 'filterElement.filterResY.baseVal = 200')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'filterResY' property of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL filterElement.filterResY.baseVal should be 400. Threw exception TypeError: 'undefined' is not an object (evaluating 'filterElement.filterResY.baseVal')
